Revert "Merge remote-tracking branch 'origin/manual_seal_mode' into feat/mantapay-v1-update"
This reverts commit faecb958b8f77c869f9abdd3f50a737efb8c1c42, reversing
changes made to 4b5c0ff17c2e9522f2a370bd546827a58455f248.
