+++ node/src/lib.rs
@@ -21,7 +21,6 @@
 extern crate alloc;
 
 mod aura_or_nimbus_consensus;
-pub mod builder;
 pub mod chain_specs;
 pub mod cli;
 pub mod client;
